+++ theicenet-cryptography-module/src/main/java/com/theicenet/cryptography/keyagreement/SRP6VerifierService.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.theicenet.cryptography.keyagreement.pake.srp.v6a;
+package com.theicenet.cryptography.keyagreement;
 
 /**
  * A SRP6VerifierService instance is a component which implements the required to generate a