+++ theicenet-cryptography-module/src/main/java/com/theicenet/cryptography/keyagreement/pake/srp/v6a/RFC5054SRP6ClientService.java
@@ -28,6 +28,7 @@ import static com.theicenet.cryptography.keyagreement.pake.srp.v6a.SRP6CommonUti
 import com.theicenet.cryptography.digest.DigestAlgorithm;
 import com.theicenet.cryptography.digest.DigestService;
 import com.theicenet.cryptography.digest.JCADigestService;
+import com.theicenet.cryptography.keyagreement.SRP6ClientService;
 import com.theicenet.cryptography.random.SecureRandomDataService;
 import java.math.BigInteger;
 import org.apache.commons.lang.Validate;