+++ theicenet-cryptography-module/src/test/java/com/theicenet/cryptography/keyagreement/pake/srp/v6a/RFC5054SRP6ServerServiceTest.java
@@ -31,6 +31,7 @@ import static org.junit.jupiter.api.Assertions.assertThrows;
 
 import com.theicenet.cryptography.digest.DigestAlgorithm;
 import com.theicenet.cryptography.digest.JCADigestService;
+import com.theicenet.cryptography.keyagreement.SRP6ServerService;
 import com.theicenet.cryptography.random.JCASecureRandomDataService;
 import com.theicenet.cryptography.test.support.HexUtil;
 import com.theicenet.cryptography.test.support.RunnerUtil;