+++ CHANGELOG.md
@@ -7,6 +7,25 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.1.2] - 2021‑03‑20
+
+### Added
+
+### Changed
+
+- SRP6 interfaces moved to package `com.theicenet.cryptography.keyagreement` [@jjfidalgo]
+- Updated README.md [@jjfidalgo]
+
+### Deprecated
+
+### Removed
+
+- SRP6 interfaces removed from package `com.theicenet.cryptography.keyagreement.pake.srp.v6a` [@jjfidalgo]
+
+### Fixed
+
+### Security
+
 ## [1.1.1] - 2021‑03‑19
 
 ### Added
@@ -94,7 +113,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Random data generation for Initialization Vector (IV) and Salt [@jjfidalgo]
 - Bitbucket pipeline with the CI & release process to Maven's Central [@jjfidalgo]
 
-[unreleased]: https://github.com/theicenet/theicenet-cryptography/compare/theicenet-cryptography-1.1.1...HEAD
+[unreleased]: https://github.com/theicenet/theicenet-cryptography/compare/theicenet-cryptography-1.1.2...HEAD
+[1.1.2]: https://github.com/theicenet/theicenet-cryptography/compare/theicenet-cryptography-1.1.1...theicenet-cryptography-1.1.2
 [1.1.1]: https://github.com/theicenet/theicenet-cryptography/compare/theicenet-cryptography-1.1.0...theicenet-cryptography-1.1.1
 [1.1.0]: https://github.com/theicenet/theicenet-cryptography/compare/theicenet-cryptography-1.0.3...theicenet-cryptography-1.1.0
 [1.0.3]: https://github.com/theicenet/theicenet-cryptography/compare/theicenet-cryptography-1.0.2...theicenet-cryptography-1.0.3